Add Gradle Enterprise Maven Extension
This adds the Gradle Enterprise Maven Extension for build scans and local build cache.
Build scans will only be published if the user is already authenticated. This is
the same behavior that is being used for the Gradle builds.
In this change, the remote build cache is not yet enabled.